src/backend/requirements.txt
requests-oauth==0.4.1
requests-oauthlib==1.3.0
requests==2.23.0

django==3.1.14
django-storages==1.9.1
django-filter==2.2.0
django-cors-headers==3.2.1
django-user-accounts==3.0.0
django-rest-auth==0.9.5
django-prettyjson==0.4.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
django-allauth==0.42.0

psycopg2-binary==2.8.4
zappa==0.48.2
uwsgi==2.0.19.1
django-anymail[mailgun]==7.0.0
botocore==1.15.34
boto3==1.12.34
python-magic==0.4.15
Pygments==2.7.4
pyaml==20.4.0
docker==4.2.0
ruamel.yaml==0.16.10
better_profanity==0.6.1